Refractor Code
The current change is a step one of a partial rewrite. This will allow commands to be easily implemented without numeorus of classes extending CommandBase. The way it works is by using reflection to invoke the method that go respond to the command that is being called. Please refer to CommandManager.class and any of the other classes to see how it works.