Code Modification
Added some caching into the mod to give it more proformance.

Changed methods to use/rely on whats in CommandBase.

Added new annotation field called "disableInProduction". This will make it so when we are pushing the mod out for production/release, we can disable buggy commands by default, but otherwise have it enabled in debug mode.

Created a new inner class called CommandStructure. This will store the annotation information (eg command name, permission node, etc) to be able to use later with how we have commands setup.

Added 3 methods in the Player class to check to see if the player is muted, banned, or jailed. Will be used for future commands and hooks.

Rewritten how the player is checked during chat, login, and logout. Also added a cleanup to remove the player from memory when they logout to prevent possible memory leakages.

Change loadFile to use String instead of BufferedReader to fix an IOException when reading json file dur to the reader being closed before it is returned. Also added writeToFile to make it possible to write to a file in a simple manner.

Added (but will likely remove) KitInventory and KitTemplate class. This will be used for managing kits.

Purged Profile.java

Added forgotten class CommonCommands to the cmdList. Will find a better method of loading commands without loading the class manually into the arraylist.

Added method "doesPermissionExExist" to check and see if it does when the mod is not in debug mode.