Cleanup annotations
Merge from upstream/master

Created and moved smite, entitycannon (previously was kittycannon but now with several entities and mainly for testing), summon (indev), fly, and god mode into FunCommands (may rename to ExtraCommands)

Changed code from upstream to now support Player.sendSuccessMessage() and Player.sendErrorMessage() which will send the message when there is an error or an successful task.

Added placeholders for tpa, tpall, tpaccept, tpreject, and warpall

Cleanup and partially finished lag command to show global information as well as per world (arguments must be the dimension numbers).

Move the check so when it is not in debug mode, it would detect the client in the preInit and if Pex/PermissionForge exist in serverStarting

Removed unused arraylist

Added placeholder for OfflinePlayer

Fix casting in Player.exec() to use ePlayer instead of "this" which caused a crash when using sudo (my mistake because i forgot to change it back)

Cleanup and added full support for fly and god mode (so far it works perfectly)

From upstream, added support for powertools (still in-dev)

In Scheduler, removed unused/empty methods, and added exec.shutdown to cancelTask

Added EntityBombTask for entitycannon

Clean up DataUtils and renamed DateUtils adding a few methods inside for supporting time.